+++ fdtool/modules/ObtainEquivalences.py
