+++ fdtool/modules/Prune.py
