+++ fdtool/modules/binaryRepr.py
